# apps/admin-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/login": {
      "post": {
        "description": "Login with username and password, returns Access Token and Refresh Token. Max 20 attempts per minute.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful, returns accessToken and refreshToken",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "eyJ...",
                    "refreshToken": "eyJ..."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid username or password"
          },
          "429": {
            "description": "Too many requests, rate limit exceeded"
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Use Refresh Token to obtain new Access Token and Refresh Token (token rotation). Old Refresh Token is immediately invalidated.",
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "eyJ...",
                    "refreshToken": "eyJ..."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Refresh Token is invalid or expired"
          }
        },
        "summary": "Refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Logout current user and revoke all valid Refresh Tokens to prevent token reuse.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "success": true
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Logout (revoke all Refresh Tokens)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "description": "Return basic info of the current logged-in user including ID, username, and roles.",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User info",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 1,
                    "username": "admin",
                    "roles": [
                      "admin"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Get current user info",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/totp/setup": {
      "post": {
        "description": "Generates a Base32 secret and otpauth:// URL, stores it in staged (disabled) state. Call POST /auth/totp/enable with a valid code from the authenticator to activate.",
        "operationId": "AuthController_totpSetup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Staged secret + otpauth URL"
          },
          "400": {
            "description": "TOTP already enabled"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Stage a TOTP secret (2FA opt-in, step 1)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/totp/enable": {
      "post": {
        "description": "Verifies one valid code against the staged secret, then sets totpEnabled=true.",
        "operationId": "AuthController_totpEnable",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "TOTP enabled"
          },
          "400": {
            "description": "Invalid code or nothing staged"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Enable TOTP (2FA opt-in, step 2)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/totp/disable": {
      "post": {
        "description": "Requires the account password or a valid TOTP code as confirmation.",
        "operationId": "AuthController_totpDisable",
        "parameters": [],
        "responses": {
          "200": {
            "description": "TOTP disabled (or was not on)"
          },
          "401": {
            "description": "Confirmation failed"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Disable TOTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/totp/verify": {
      "post": {
        "description": "Re-validates username+password, verifies the 6-digit TOTP code, and issues accessToken/refreshToken. Called after /auth/login returned { totpRequired: true }.",
        "operationId": "AuthController_totpVerifyLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens issued",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "accessToken": "eyJ...",
                    "refreshToken": "eyJ..."
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials or TOTP code"
          }
        },
        "summary": "Complete TOTP login (second factor)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/sessions": {
      "get": {
        "description": "One row per active (non-revoked, non-expired) refresh token. The row matching the sid claim of the caller's access token is flagged current=true.",
        "operationId": "AuthController_listSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session list"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "List my active login sessions",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/sessions/{id}": {
      "delete": {
        "description": "Marks the given refresh token row revoked (DR-04 semantics — the refresh JWT becomes unusable immediately). Ownership is enforced; foreign ids get 401.",
        "operationId": "AuthController_revokeSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session revoked"
          },
          "401": {
            "description": "Session not found or not yours"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Revoke one of my sessions",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/sessions/revoke-others": {
      "post": {
        "description": "Bulk remote-logout of other devices. Without a sid claim this degenerates to revoking ALL sessions (fail-safe, never keeps unknown sessions).",
        "operationId": "AuthController_revokeOtherSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Revocation count"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "Revoke all my sessions except the current one",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "AutoFlow Admin API",
    "description": "AutoCodeFlow Admin API — exported by test/openapi-export.e2e-spec.ts (ARCH-23). Single source of truth for generated frontend types (apps/admin-web gen:api-types) and the CI api-types-drift check. Regenerate with `npm run swagger:export` after any API contract change.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3105",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "password123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TotpCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6 位 TOTP 动态验证码"
          }
        },
        "required": [
          "code"
        ]
      },
      "TotpVerifyDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "password123"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "description": "6 位 TOTP 动态验证码"
          }
        },
        "required": [
          "username",
          "password",
          "code"
        ]
      },
...
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}